+++ src/App.js
@@ -1,21 +1,23 @@
 import React, {Component} from 'react';
-import './assets/css/index.css'
+import './assets/css/main.css'
 import {BrowserRouter as Router, Route} from 'react-router-dom';
-import Navigation from './components/Navigation';
-import Portfolio from "./components/Portfolio";
-import GeneratePDF from "./components/GeneratePDF";
+import Navigation from './components/Header/Navigation';
+import Portfolio from "./components/Main/Main";
+import Footer from "./components/Footer/footer";
 
 class App extends Component {
     render() {
         return (
             <Router key={'router'}>
-                <div>
-                    {<Navigation key={'super-navigation'}/>}
-                </div>
-                <div style={{marginTop: "4rem"}}>
+                <header>
+                    {<Navigation key={'navigation'}/>}
+                </header>
+                <main key={'main'}>
                     <Route path='/' exact component={Portfolio}/>
-                    <Route path='/resume' exact component={GeneratePDF}/>
-                </div>
+                </main>
+                <footer>
+                    {<Footer key={'footer'}/>}
+                </footer>
             </Router>
         );
     }
