+++ src/archive/main.scss
@@ -1,0 +1,5 @@
+/* General Layout */
+
+body {
+    background-color: white;
+}
