+++ src/archive/portfolio.css
@@ -1,5 +1,5 @@
 body {
-    background-color: #8EC38D;
+    background-color: rgba(255, 255, 255, 0.884);
 }
 
 .bg-light {
