+++ src/assets/css/main.css
@@ -6,10 +6,7 @@
 body {
     margin: 0;
     font-family: "Roboto Slab", serif;
-}
-
-code {
-    font-family: "Roboto Slab", serif;
+    background-color: rgba(255, 255, 255, 0.884);
 }
 
 #brand {
