--- src/assets/scss/portfolio.scss
@@ -1,2 +1,0 @@
-@import "_partials/smoky_effect.scss";
-@import "../css/portfolio.css";
