+++ src/components/Main/sections/about.js
@@ -2,7 +2,7 @@ import React from "react";
 
 function about() {
     return (
-        <div id="about" className="container-fluid">
+        <section id="about" className="container-fluid">
             <div className="anchor">
                 <div className="card">
                     <div className="card-header">
@@ -29,7 +29,7 @@ function about() {
                     </div>
                 </div>
             </div>
-        </div>
+        </section>
     )
 }
